backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.29.0
bcrypt==4.3.0
certifi==2025.4.26
cffi==1.17.1
click==8.1.8
cryptography==44.0.3
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.1.1
fastapi==0.115.12
fastapi-cli==0.0.7
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
orjson==3.10.18
packaging==25.0
passlib==1.7.4
psycopg2==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.7.1
pydantic-core==2.18.2
pydantic-settings==2.9.1
Pygments==2.19.1
pytest==8.3.5
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
PyYAML==6.0.2
rich==14.0.0
rich-toolkit==0.14.5
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
sqlmodel==0.0.24
starlette==0.46.2
typer==0.15.3
typing-extensions==4.13.2
typing-inspection==0.4.0
ujson==5.10.0
uvicorn==0.30.1
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
black==23.9.1
isort==5.12.0